Texture pack streaming, Ooccoo quick-use, guide reader off the frame path
Updates aurora to upstream (24 commits), replacing the fork's synchronous
texture replacement with upstream's streaming loader: worker threads,
thumbnail preloads, and a per-frame GPU publish cap. On an 8.2 GB pack the
intro's memory-availability collapse drops from 1101 MB to 297 MB and the
kernel's low-memory reclaim stops firing entirely.

Android glue now subclasses dev.encounter.aurora.AuroraSurface instead of
patching SDL's SDLSurface.java, so no fork patch to SDL's Java remains.
AIInitDMA's stub follows upstream's widened uintptr_t signature.

Adds the Ooccoo quick-use button to the map, moves guide image decoding off
the render thread, and drops duplicate images at guide import (converter
version 10 -- existing libraries reconvert in place, no re-download).

+++ src/dusk/stubs.cpp
@@ -908,7 +908,11 @@ void AIInit(u8* stack) {
     // In a real scenario, it would set up the audio interface and prepare it for use.
 }
 
-void AIInitDMA(u32 start_addr, u32 length) {
+// uintptr_t, not u32: upstream aurora widened this on TARGET_PC so a 64-bit
+// host address survives the call (dolphin/ai.h guards the two spellings). The
+// stub has to match that declaration or it defines a different symbol and
+// JASAiCtrl's caller goes unresolved at link time.
+void AIInitDMA(uintptr_t start_addr, u32 length) {
     STUB_LOG();
 }
